moto_swap: Refactor and fix the driver
moto_mm, moto_swap: fix field name, race, duplicate def, and missing type

Fix A (moto_mmap_fault): vm_fault has android_vendor_data1, not
android_oem_data1. Replace all references with the correct field
name using packing macros to store old_ra_pages and max_ra_pages in a
single u64.

Fix B (moto_mmap_fault): Document the version gate's purpose — it detects
availability of the android_vh_tune_mmap_readaround hook. No functional change.

Fix C (moto_mmap_fault): The legacy filemap_fault hooks can race when
two threads fault the same file concurrently, causing the original
ra_pages value to be lost permanently. Add an ownership flag (bit 32)
so only the first fault saves the original value and only that fault
restores it. Use READ_ONCE/WRITE_ONCE for compiler barriers.

Fix D (hybridswap_internal.h): Remove the non-static duplicate
definition of hybridswap_fetch_zram_pagefault in the
!CONFIG_HYBRIDSWAP_CORE fallback block. Only the static inline
version is needed.

Fix E (hybridswap_eswap.c): Add missing ssize_t return type to
meminfo_show(). Implicit int is an error with -Werror=implicit-int.

moto_mm: drop redundant module

Remove drivers/moto_mm/ entirely. On 5.4 this module is 95% dead code:
- drain_all_pages_bypass hook: only for 5.10-5.15, not compiled
- mm_info alloc warning hooks: only for >= 5.10, not compiled
- proc interface (/proc/moto_mm/): only for >= 5.10, not created
- tune_inactive_ratio: the only working feature, but redundant with
  moto_swap which already handles aggressive anon reclaim via
  swappiness=200 and SCAN_ANON scan type

The kernel already provides CONFIG_MMAP_READAROUND_LIMIT in
mm/filemap.c for readahead limiting, making moto_mmap_fault
equally redundant.

moto_swap: fix critical, high, and medium severity bugs

Critical:
- hybridswap_core_init(): return ret on bind failure instead of
  hardcoded 0. Caller was proceeding with broken state.
- __GFP_NOFAIL + GFP_ATOMIC: replaced with GFP_KERNEL to avoid
  infinite retry loop that can soft-lockup the CPU.
- free_pages(0, 1) on NULL buffer: add NULL check in zcomp_strm_free()
  to prevent freeing physical page 0. Fixed in all 3 zram copies.

High:
- hybridswap_loop_device_store(): propagate hybridswap_core_init()
  error to userspace instead of always returning success.
- bitmap_sz: change from unsigned int to size_t to prevent truncation
  on large backing devices. Fixed in all 3 zram copies.

Medium:
- sched_setscheduler_nocheck(): log error if swapd priority change
  fails instead of silently ignoring it.
- 14 inline functions in hybridswap_swapd.c: add missing static keyword
  to prevent global symbol namespace pollution.
- hybridswap_core_enable_show(): clamp snprintf return to PAGE_SIZE-1
  to prevent sysfs buffer over-read on truncation.

moto_swap: use /proc/sys/vm/swappiness instead of hardcoded values

The tune_swappiness hook was hardcoding swappiness=200 for mswapd
and 60 for direct reclaim. This overrides the admin's
/proc/sys/vm/swappiness setting.

swappiness=200 is particularly harmful: it sets anon_prio=200,
file_prio=0 (200-200), completely starving the page cache. The
tune_scan_type hook already forces SCAN_ANON for mswapd, making
the swappiness override redundant.

Fix: all three paths (kswapd, mswapd, direct reclaim) now read
vm_swappiness from /proc/sys/vm/swappiness. The only exception
is mswapd when swap is low, where swappiness is set to 0 to
prevent further swapping.

moto_swap: remove dead alloc_pages_slowpath_hook

The alloc_pages_slowpath_hook function was defined in
hybridswap_swapd.c and declared in hybridswap_internal.h, but
never registered — no DECLARE_HOOK in kernel headers, no
trace_android_vh_* call in mm/. The rmqueue_hook (which IS
registered) does the exact same thing: wake_all_swapd().

moto_swap: add MODULE_ALIAS for zram

The moto_swap module bundles its own zram driver but the module
is named moto_swap.ko, not zram.ko. Kernel managers and modprobe
that look for 'zram' won't find it. Add MODULE_ALIAS so it's
discoverable as zram.

+++ drivers/moto_swap/zram-5.10/zram_drv.c
@@ -392,7 +392,7 @@ static ssize_t backing_dev_store(struct device *dev,
 	struct file *backing_dev = NULL;
 	struct inode *inode;
 	struct address_space *mapping;
-	unsigned int bitmap_sz, old_block_size = 0;
+	size_t bitmap_sz, old_block_size = 0;
 	unsigned long nr_pages, *bitmap = NULL;
 	struct block_device *bdev = NULL;
 	int err;
